// src/config/enums.ts
enum URLS {
  FACEBOOK = "https://www.facebook.com/alfab_electric",
  INSTAGRAM = "https://www.instagram.com/alfab_electric/",
  LINKEDIN = "https://www.linkedin.com/in/alfab-electric-3a4b1b28b/",
  TIKTOK = "https://www.tiktok.com/@alfab.electric?_t=8f3uKXbYOxy&_r=1",
  WHATSAPP = "https://wa.me/593998031971",
  BROCHURE = "./public/Brochure.pdf",
}

enum ROUTES {
  INICIO = "/",
  NOSOTROS = "/nosotros",
  SERVICIOS = "/servicios",
  PROYECTOS = "/proyectos",
  CONTACTENOS = "/contactenos",
}

export default URLS;
export { ROUTES };